Fix plain-text password reset link from being wiped out.
The issue is the password reset link is wrapped around brackets:
eg. <hxxp://example.com/wp-login.php?action=rp...>

Our HTML to plain-text parser thinks the password reset link is a HTML tag.
When the parser encounters invalid HTML tags (as in the case of the
password reset link), it removes it.

This commit removes the brackets from the password reset link before
our HTML to plain-text conversion takes place, which should address this
issue for plain-text email users.